feat: verify if vsce pat is valid or not
