+++ lib/verify.js
@@ -1,5 +1,5 @@
 const verifyPkg = require('./verify-pkg');
-const setAuth = require('./set-auth');
+const verifyAuth = require('./verify-auth');
 const getPkg = require('read-pkg-up');
 const SemanticReleaseError = require('@semantic-release/error');
 
@@ -12,10 +12,6 @@ module.exports = async logger => {
     );
   }
   verifyPkg(packageJson);
-  await setAuth(logger);
-  try {
-    // TODO: Verify output
-  } catch (err) {
-    throw new SemanticReleaseError('Invalid vsce token.', 'EINVALIDVSCETOKEN');
-  }
+
+  verifyAuth(logger);
 };
